+++ src/chains/HyperEvm/Common.sol
@@ -10,6 +10,11 @@ import {ISettlerActions} from "../../ISettlerActions.sol";
 import {ISignatureTransfer} from "@permit2/interfaces/ISignatureTransfer.sol";
 import {revertUnknownForkId} from "../../core/SettlerErrors.sol";
 
+import {IUniswapV3Callback} from "../../core/univ3forks/UniswapV3.sol";
+import {kittenSwapFactory, kittenSwapInitHash, kittenSwapForkId} from "../../core/univ3forks/KittenSwap.sol";
+import {hybraFactory, hybraInitHash, hybraForkId} from "../../core/univ3forks/Hybra.sol";
+import {hyperSwapFactory, hyperSwapInitHash, hyperSwapForkId, IHyperswapV3SwapCallback} from "../../core/univ3forks/HyperSwap.sol";
+
 // Solidity inheritance is stupid
 import {SettlerAbstract} from "../../SettlerAbstract.sol";
 
@@ -39,6 +44,20 @@ abstract contract HyperEvmMixin is FreeMemory, SettlerBase {
         override
         returns (address factory, bytes32 initHash, uint32 callbackSelector)
     {
-        revertUnknownForkId(forkId);
+        if (forkId == kittenSwapForkId) {
+            factory = kittenSwapFactory;
+            initHash = kittenSwapInitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == hybraForkId) {
+            factory = hybraFactory;
+            initHash = hybraInitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == hyperSwapForkId) {
+            factory = hyperSwapFactory;
+            initHash = hyperSwapInitHash;
+            callbackSelector = uint32(IHyperswapV3SwapCallback.hyperswapV3SwapCallback.selector);
+        } else {
+            revertUnknownForkId(forkId);
+        }
     }
 }